+++ src/autotesting/CMakeLists.txt
@@ -1,6 +1,5 @@
 set(AutoTesting_SOURCES
   AutowiringEnclosure.h
-  AutowiringEnclosure.cpp
   gtest-all-guard.h
   gtest-all-guard.cpp
 )
